Fixed problem with commited transactions, currently, transactions are replaced with transactions from accounting after import.
Still problem with commited transactions and autobooking and verification import, verifications are imported twice, even thought they are commited.

This needs to be reworked so that external transactions are kept separate from booked transactions.

Fixed bokslut, needed a crude hack to get the dates right. Date handling should be formalised so that the same timestamp is used for all dates, e.g, time of day.